% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting_format.R
\name{accounting_format}
\alias{accounting_format}
\title{accounting_format}
\usage{
accounting_format(x, digits = 0, zero.print = ".")
}
\arguments{
\item{x}{is the object to apply the accounting format to}

\item{digits}{is the number of digits to use in printing the statement}

\item{zero.print}{is format to use for nil}
}
\value{
this function the modified x
}
\description{
this function prints number in a nice accounting format
}
