% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\name{annuity}
\alias{annuity}
\title{annuity}
\usage{
annuity(x, output)
}
\arguments{
\item{x}{annuity object name}

\item{output}{is object containing all data}
}
\value{
this function returns a text file 'output' updated for annuity
}
\description{
function that projects the debt in the financial statements
}
\seealso{
Other cash flow: 
\code{\link{cash_account}()}
}
\concept{cash flow}
