% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_class.R
\docType{class}
\name{annuity_class-class}
\alias{annuity_class-class}
\alias{annuity_class}
\title{annuity_class}
\arguments{
\item{annual_notional}{= numeric}

\item{minimum_age_payment}{= numeric}

\item{maximum_age_payment}{= numeric}

\item{currency}{= character}

\item{inflation_indexation}{= logical}

\item{tax_rate}{= numeric}

\item{taxpayer}{= character}

\item{metrics}{= list}

\item{statement}{= "list}
}
\description{
annuity_class
}
\seealso{
Other class: 
\code{\link{cash_account_class-class}},
\code{\link{expenses_class-class}},
\code{\link{pension_class-class}}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{class}
