% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_account.R
\name{cash_account}
\alias{cash_account}
\title{cash_account}
\usage{
cash_account(x, output)
}
\arguments{
\item{x}{annuity object name}

\item{output}{is object containing all data}
}
\value{
this function returns a text file 'output' updated for cash account
}
\description{
function that projects the debt in the financial statements
}
\seealso{
Other cash flow: 
\code{\link{annuity}()}
}
\concept{cash flow}
