% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_account_class.R
\docType{class}
\name{cash_account_class-class}
\alias{cash_account_class-class}
\alias{cash_account_class}
\title{cash_account_class}
\arguments{
\item{GBP}{"numeric"}

\item{CAD}{= "numeric"}

\item{USD}{= "numeric"}

\item{EUR}{= "numeric"}

\item{KRW}{= "numeric"}

\item{CHF}{= "numeric"}

\item{metrics}{= "list"}

\item{statement}{= "list"}
}
\description{
cash_account_class
}
\seealso{
Other class: 
\code{\link{annuity_class-class}},
\code{\link{expenses_class-class}},
\code{\link{pension_class-class}}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{class}
