% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_flow.R
\name{cash_flow}
\alias{cash_flow}
\title{cash_flow}
\usage{
cash_flow(x, output)
}
\arguments{
\item{x}{is the object to apply the financial function to}

\item{output}{is object containing all data}
}
\value{
output with a modified statement
}
\description{
this function applies the different financial cash flow projection function depending on the class of 'x'
}
