% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.R
\name{debt}
\alias{debt}
\title{debt}
\usage{
debt(x, output)
}
\arguments{
\item{x}{is the debt object}

\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a output with the statement modified by debt x
}
\description{
function that projects the debt in the financial statements
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
