% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_class.R
\docType{class}
\name{debt_class-class}
\alias{debt_class-class}
\alias{debt_class}
\title{debt_class}
\arguments{
\item{balance}{amount left to pay}

\item{repayment_cash_annual}{annual repayment amount (optional if repayment_percentage_annual is used instead)}

\item{repayment_percentage_annual}{percentage of balance to repay (optional is repayment_percentage_annual is used instead)}

\item{interest_rate}{annual interest rate e.g.: 1\% is entered as 0.01}

\item{promo_interest_rate_end_date}{end date of temporary promotional annual interest rate (optional - used in conjunction with promo_interest_rate)}

\item{promo_interest_rate}{temporary annual interest rate during promotion (optional - used in conjunction with promo_interest_rate_end_date)}

\item{currency}{currency of the loan}

\item{metrics}{leave unused}

\item{statement}{leave unused}
}
\value{
a class for debt object
}
\description{
class definition for debt object
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
