% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment}
\alias{employment}
\title{employment}
\usage{
employment(x, output)
}
\arguments{
\item{x}{is the employment object}

\item{output}{is the only parameter 'input' needed.}
}
\value{
this function returns a output with the statement modified by employment x
}
\description{
function that projects the employment in the financial statements
}
