% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_class.R
\docType{class}
\name{employment_class-class}
\alias{employment_class-class}
\alias{employment_class}
\title{employment_class}
\arguments{
\item{retirement_age}{= "numeric",}

\item{yearly_salary}{= "numeric",}

\item{currency}{= "character",}

\item{taxpayer}{= "character",}

\item{tax_rate}{= "numeric",}

\item{metrics}{= "list",}

\item{statement}{= "list"}
}
\description{
employment class
}
