% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenses.R
\name{expenses}
\alias{expenses}
\title{expenses}
\usage{
expenses(x, output)
}
\arguments{
\item{x}{expense object}

\item{output}{all in memory}
}
\description{
create cash flow for expenses
}
\seealso{
Other cash flow function: 
\code{\link{pension}()},
\code{\link{property}()}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{cash flow function}
