% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenses_class.R
\docType{class}
\name{expenses_class-class}
\alias{expenses_class-class}
\alias{expenses_class}
\title{expenses_class}
\arguments{
\item{currency}{= "character"}

\item{expenses}{= "list"}

\item{metrics}{= "list"}

\item{statement}{= "list"}
}
\value{
an object of expenses class
}
\description{
expenses_class
}
\seealso{
Other class: 
\code{\link{annuity_class-class}},
\code{\link{cash_account_class-class}},
\code{\link{pension_class-class}}
}
\concept{class}
