% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_portfolio_class.R
\name{find_portfolio_class}
\alias{find_portfolio_class}
\title{find_portfolio_class}
\usage{
find_portfolio_class(portfolio)
}
\arguments{
\item{portfolio}{this is the list of all financial information (e.g. asset and liability)}
}
\value{
this function returns the class of 'portfolio'
}
\description{
function that test and find the names for a classes in the portfolio - function input : function(portfolio, class) where portfolio is a list and class is a string e.g.: find_portfolio_class(portfolio = output$Portfolio$portfolio, class = "employment_class")
}
