% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FO.R
\name{FO}
\alias{FO}
\title{FO}
\usage{
FO(input)
}
\arguments{
\item{input}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a text file 'output.txt'
}
\description{
family office wrapping function. This function calculate the financial statement projections.
}
