% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables.R
\name{format_tables}
\alias{format_tables}
\title{format_tables}
\usage{
format_tables(output)
}
\arguments{
\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a text file 'output'
}
\description{
format tables with Excel like accounting number format
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
