% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_table.R
\name{formatted_table}
\alias{formatted_table}
\title{formatted_table}
\usage{
formatted_table(data, currency, interval = 3, mark = ",", digits = 0)
}
\arguments{
\item{data}{table to format}

\item{currency}{three letter currency code (character string) e.g.: "USD", "GBP", "KRW", "CHF", "EUR", "CAD"}

\item{interval}{number grouping e.g: 1,000,000 (normally 3, Korean 4), default = 3}

\item{mark}{number group separator (normally ",", French " " i.e.: space), default = ","}

\item{digits}{number of decimals to keep, default = "0"}
}
\value{
a formatted table
}
\description{
function that format tables in accounting format
}
