% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\name{fx}
\alias{fx}
\title{fx}
\usage{
fx(output)
}
\arguments{
\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a text file 'output.txt'
}
\description{
function that rebase currency
}
