% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_metrics.R, R/global_metrics_function.R
\name{global_metrics}
\alias{global_metrics}
\title{global_metrics}
\usage{
global_metrics(output)

global_metrics(output)
}
\arguments{
\item{output}{all in memory}
}
\value{
this function returns a text file 'output'

output with updated global metrics
}
\description{
function that calculates metrics from the financial statements

produces all the global metrics in the financial statements
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))

person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
