% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_library_function.R
\name{import_library_function}
\alias{import_library_function}
\title{import_library_function}
\usage{
import_library_function()
}
\value{
nothing, function are exported to global environment
}
\description{
this function applies the different financial cash flow projection function depending on the class of 'x'
}
