% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_adjustment.R
\name{inflation_adjustment}
\alias{inflation_adjustment}
\title{inflation_adjustment}
\usage{
inflation_adjustment(amount, currency, output)
}
\arguments{
\item{amount}{is the amount to adjust for inflation}

\item{currency}{is used to select the inflation to use}

\item{output}{contains all data.}
}
\value{
this function returns the amount modified for inflation
}
\description{
function that adjust projection monetary amount for inflation in the financial statements
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
