% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_function.R
\name{input_function}
\alias{input_function}
\title{input_function}
\usage{
input_function(input, output)
}
\arguments{
\item{input}{is the only parameter 'input' needed is the input files name.}

\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a text file 'output.txt'
}
\description{
family office wrapping function. This function calculate the financial statement projections.
}
