% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{iterate}
\usage{
iterate(output)
}
\arguments{
\item{output}{is the only parameter needed.}
}
\value{
this function returns a text file 'output'
}
\description{
iterate over time and portfolio steps
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
