% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linebreaks.R
\name{linebreaks}
\alias{linebreaks}
\title{linebreaks}
\usage{
linebreaks(n)
}
\arguments{
\item{n}{number of lines to skip}
}
\value{
this function returns a text file 'output.txt'
}
\description{
skip multiple line in HTML
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{user interface}
