% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_portfolio_all.R
\name{make_portfolio_all}
\alias{make_portfolio_all}
\title{make_portfolio_all}
\usage{
make_portfolio_all(user)
}
\arguments{
\item{user}{identity of the user.}
}
\value{
this function returns the
}
\description{
function that test and find the names for a classes in the portfolio - function input : function(portfolio, class) where portfolio is a list and class is a string e.g.: find_portfolio_class(portfolio = output$Portfolio$portfolio, class = "employment_class")
}
