% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_total.R
\name{metric_total}
\alias{metric_total}
\title{metric_total}
\usage{
metric_total(x)
}
\arguments{
\item{x}{is the only parameter needed is the part the financial statements to summarise.}
}
\value{
this function returns a text file 'output'
}
\description{
function that calculates sum of rows of financial statements to summarise the information and adds the metric to the statement
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
