% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage.R
\name{mortgage}
\alias{mortgage}
\title{mortgage}
\usage{
mortgage(principal, annual_rate, frequency, mortgage_term_in_year)
}
\arguments{
\item{principal}{principal amount, numeric}

\item{annual_rate}{numeric e.g.: 1\% is 0.01}

\item{frequency}{choice of "years", "quarters", "months", "weeks", "days"}

\item{mortgage_term_in_year}{number of years on mortgage}
}
\value{
mortgage payment: amortisation interest and principal
}
\description{
mortgage function
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
