% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pension.R
\name{pension}
\alias{pension}
\title{pension}
\usage{
pension(x, output)
}
\arguments{
\item{x}{the pension object}

\item{output}{all in memory}
}
\value{
output with updated pension
}
\description{
produce cash flow for pension object
}
\seealso{
Other cash flow function: 
\code{\link{expenses}()},
\code{\link{property}()}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{cash flow function}
