% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pension_class.R
\docType{class}
\name{pension_class-class}
\alias{pension_class-class}
\alias{pension_class}
\title{pension_class}
\arguments{
\item{value}{numeric}

\item{minimum_age_withdrawal}{numeric}

\item{currency}{character}

\item{cash_withdrawal}{numeric}

\item{percentage_withdrawal}{numeric}

\item{tax_rate}{numeric}

\item{taxpayer}{character}

\item{metrics}{list}

\item{statement}{list}
}
\description{
pension_class
}
\seealso{
Other class: 
\code{\link{annuity_class-class}},
\code{\link{cash_account_class-class}},
\code{\link{expenses_class-class}}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{class}
