% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picture_choice.R
\name{picture_choice}
\alias{picture_choice}
\title{picture_choice}
\usage{
picture_choice()
}
\description{
randomise the choice of picture for graphical user interface background
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))S
}
