% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FO_2.R
\name{plot_FO_2}
\alias{plot_FO_2}
\title{plot_FO_2}
\usage{
plot_FO_2(FO, input)
}
\arguments{
\item{FO}{object of family office to plot}

\item{input}{input parameters to plot}
}
\description{
plotting function for statements
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
