% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_to_statement.R
\name{portfolio_to_statement}
\alias{portfolio_to_statement}
\title{portfolio_to_statement}
\usage{
portfolio_to_statement(output, portfolio_name, date)
}
\arguments{
\item{output}{is the only parameter 'input' needed is the input files name.}

\item{portfolio_name}{is the only parameter 'input' needed is the input files name.}

\item{date}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns 'output' with a modified statement
}
\description{
function that calculates the impact of a portfolio object 'portfolio_name' on the financial statement for a specifc date
}
