% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cash_flow.R
\name{process_cash_flow}
\alias{process_cash_flow}
\title{process_cash_flow}
\usage{
process_cash_flow(input)
}
\arguments{
\item{input}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a text file 'output.txt'
}
\description{
This function calls each of the function calls all the other function needed to creates financial statement.
}
