% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFile2.R
\name{processFile2}
\alias{processFile2}
\title{processFile2}
\usage{
processFile2(user, class)
}
\arguments{
\item{user}{name of user 'character'}

\item{class}{name of the class}
}
\description{
return the sub-portfolio for a class for a given user
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
