% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_initialisation.R
\name{program_initialisation}
\alias{program_initialisation}
\title{program_initialisation}
\usage{
program_initialisation()
}
\value{
this function returns a text file 'output.txt'
}
\description{
initialise values like tax rates.
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
