% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property}
\alias{property}
\title{property}
\usage{
property(x, output)
}
\arguments{
\item{x}{property object}

\item{output}{all in memory}
}
\description{
Property function
}
\seealso{
Other cash flow function: 
\code{\link{expenses}()},
\code{\link{pension}()}
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
\concept{cash flow function}
