% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_class.R
\docType{class}
\name{property_class-class}
\alias{property_class-class}
\alias{property_class}
\title{property_class}
\arguments{
\item{property_value}{"numeric" e.g.: '895000'}

\item{building_purchase_price}{"numeric" e.g.: '450000'}

\item{currency}{"character", e.g.: '"CAD"'}

\item{mortgage_type}{"character" e.g.: 'fixed' or 'variable'}

\item{mortgage_term}{"character", e.g.: '25'}

\item{mortgage_spread}{"numeric" e.g.: -0.01}

\item{mortgage_interest_rate_if_fixed}{"numeric", e.g.: '0.0395' = 3.95\% per annum}

\item{mortgage_initial_rate_maturity}{"character", e.g.: '"2011-09-24"'}

\item{mortgage_holiday_end_date}{"character", date at which mortgage holiday ends e.g.: '"2022-01-29"'}

\item{mortgage_loan_value}{"numeric", e.g.: '600000'}

\item{HELOC_balance}{"numeric", e.g.: '30000'}

\item{HELOC_interest_rate_spread}{"numeric", e.g.: '0.0595' = 5.95\% per annum}

\item{HELOC_max_loan_to_value}{"numeric", e.g. '0.8'}

\item{refinancing_balance}{"numeric", e.g.: '50000'}

\item{refinance_interest_deductable}{"logical", e.g.: 'TRUE' or 'FALSE'}

\item{capital_repayment_mortgage}{"logical", e.g.: if 'TRUE' then capital repayment are calculated}

\item{capital_repayment_proportion}{"numeric", e.g.: usually 1 for full capital repayment, not used if not capital repayment, 0 for interest only, any real value can be used}

\item{home_equity_line_of_credit}{"logical", e.g.: if 'TRUE' then when mortgage reaches 80\% loan to value, withdrawal are taken}

\item{refinancing_loan_to_value_max}{"numeric", e.g.: '0.8' for 80\% loan to value}

\item{refinancing}{"logical", e.g.: 'TRUE' if true then refinancing happens when hitting 'refinancing_loan_to_value_max'}

\item{taxable}{"logical", e.g.: 'TRUE' if true then tax applies to rental income}

\item{tax_rate}{"numeric", e.g.: '0.0595' = 5.95\% per annum}

\item{taxpayer}{"character", e.g.: 'FREDERIC' tax payer name}

\item{depreciation}{"numeric", e.g.: '0.04', 4\% per annum}

\item{cumulative_losses}{"numeric", e.g.; '50000', money spent on property which can be offset against taxable income}

\item{occupancy}{"numeric", e.g.: '0.95' means the property is empty (free of rent) 5\% of the time}

\item{tenant_finders_fee_percentage_of_rent}{"numeric", e.g.: '0.03' meaning 3\% of rent in addition to tenant finders fee if applicable}

\item{property_management_percentage_of_rent}{"numeric"}

\item{inventory_fee_annual}{"numeric"}

\item{contract_fee_annual}{"numeric"}

\item{rent_annual}{"numeric"}

\item{initial_rent_annual_end_date}{"character"}

\item{initial_rent_annual}{"numeric"}

\item{boiler_insurance_annual}{"numeric"}

\item{ground_rent_incl._building_insurance_annual}{"numeric"}

\item{insurance_premium_annual}{"numeric"}

\item{property_maintenance_annual}{"numeric"}

\item{accountant_fee_annual}{"numeric"}

\item{solicitor_fee_annual}{"numeric"}

\item{municipal_and_school_tax_annual}{"numeric"}

\item{borough_charges}{"numeric"}

\item{broadband_annual}{"numeric"}

\item{water_tenant_annual}{"numeric"}

\item{water_owner_annual}{"numeric"}

\item{heathing_tenant_annual}{"numeric"}

\item{heathing_owner_annual}{"numeric"}

\item{electricity_tenant_annual}{"numeric"}

\item{electricity_owner_annual}{"numeric"}

\item{metrics}{"list"}

\item{statement}{"list"}
}
\value{
this function returns a property class object
}
\description{
function that set the class for property
}
