% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejig_output.R
\name{rejig_output}
\alias{rejig_output}
\title{rejig_output}
\usage{
rejig_output(output)
}
\arguments{
\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns the output with modified statements
}
\description{
funcion that calculates the totals and subtotals for the financial statements
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
