% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_premium_adjustment.R
\name{risk_premium_adjustment}
\alias{risk_premium_adjustment}
\title{risk_premium_adjustment}
\usage{
risk_premium_adjustment(amount, asset_class, output)
}
\arguments{
\item{amount}{value in currency}

\item{asset_class}{property or}

\item{output}{all in memory}
}
\value{
cash_flow (increased by risk premium)
}
\description{
risk_premium_adjustment
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
