% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\title{statement}
\usage{
statement(output)
}
\arguments{
\item{output}{is the only parameter 'input' needed is the input files name.}
}
\value{
this function returns a 'output'
}
\description{
function that creates the statement
}
