% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_total_program.R
\name{statement_total_program}
\alias{statement_total_program}
\title{statement_total_program}
\usage{
statement_total_program(X, reporting_date)
}
\arguments{
\item{X}{statement to apply the sub-total}

\item{reporting_date}{date of the projection}
}
\value{
this function returns X modified
}
\description{
program that sums up all subsection of statements for each reporting date; function that calculate the sub total for of a statement say all debt to total debt for a specified reporting date
}
\author{
person(given = "frederic", family = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
