% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_calculations.R
\name{tax_calculations}
\alias{tax_calculations}
\title{tax_calculations}
\usage{
tax_calculations(x, output, income)
}
\arguments{
\item{x}{is an item in the portfolio}

\item{output}{contains all data}

\item{income}{is the income}
}
\value{
output with a modified income tax
}
\description{
function that calculate income tax for the statement
}
