% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxpayer.R
\name{taxpayer}
\alias{taxpayer}
\title{taxpayer}
\usage{
taxpayer(user, n)
}
\arguments{
\item{user}{username}

\item{n}{order of the taxpayer for which to return the name e.g.: '1' or '2'}
}
\description{
check the input file associated with the user for taxpayer n's name
}
\author{
person(given = "frederic", given = "elcherif", email = "f.elcherif@gmail.com", role = c("aut", "cre"))
}
