% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_choice.R
\name{user_choices}
\alias{user_choices}
\title{user_choices}
\usage{
user_choices(user)
}
\arguments{
\item{user}{this is the name of the user}
}
\value{
this function returns a text file 'output.txt'
}
\description{
function that returns the files accessible to a user in the app folder
}
